Архитектурные изменения
