+++ .gitignore
@@ -1,2 +1,4 @@
 .idea
 __pycache__/
+.venv
+pyproject.lock
