+++ models/Item.py
@@ -1,0 +1,5 @@
+from pydantic import BaseModel
+
+
+class Item(BaseModel):
+    text: str
