+++ pyproject.toml
@@ -1,0 +1,21 @@
+[project]
+name = "se-fastapi-example"
+version = "0.1.0"
+description = ""
+authors = [
+    {name = "Viclor3"}
+]
+readme = "README.md"
+requires-python = ">=3.11"
+dependencies = [
+    "fastapi (>=0.115.8,<0.116.0)",
+    "transformers (>=4.49.0,<5.0.0)",
+    "fastapi-cli (>=0.0.7,<0.0.8)",
+    "tensorflow (>=2.18.0,<3.0.0)",
+    "tf-keras (>=2.18.0,<3.0.0)"
+]
+
+
+[build-system]
+requires = ["poetry-core>=2.0.0,<3.0.0"]
+build-backend = "poetry.core.masonry.api"
